Api42/requirements.txt
django==4.2.10
psycopg[binary] 
djangorestframework==3.15.2
djangorestframework-simplejwt
django-cors-headers
django-extensions
python-decouple
django-environ
python-dotenv
requests
Djoser
uvicorn
pyotp
qrcode[pil]
requests
pillow
twilio
environ
daphne
httpx
docker
channels
channels-redis
selenium

